renamed RealTestPID to TestPID
