+++ src/main/java/frc/robot/Robot.java
@@ -49,7 +49,7 @@ public class Robot extends TimedRobot {
         SmartDashboard.putData("Auto mode", autonomousChooser);
         SmartDashboard.putData("CalibrateDistance", new CalibrateDistance(oi.driver::getAButton));
         dbc.addNumber("limelight distance", limelight::getDistance);
-        SmartDashboard.putData("Test PID vision", new RealTestPID());
+        SmartDashboard.putData("Test PID vision", new TestPID());
         SmartDashboard.putData("clearPreferences", Commands.clearPreferences());
     }
